Support uploading files

+++ lib/api/api_request.dart
@@ -1,5 +1,7 @@
 import 'dart:convert';
+import 'dart:io';
 
+import 'package:path/path.dart';
 import 'package:http/http.dart' as http;
 import 'package:openwykop/api/api_properties.dart';
 
@@ -19,12 +21,14 @@ class ApiRequest {
   Map<String, String>? queryParameters;
   RequestMethod method;
   Map<String, dynamic> bodyData;
+  Map<String, File> files;
 
   ApiRequest(this.apiProperties, this.pathSegments, {
     this.queryParameters,
     Map<String, String>? headers,
     this.method = RequestMethod.GET,
     this.bodyData = const {},
+    this.files = const {},
   }) {
     initializeHeaders(headers);
   }
@@ -66,23 +70,35 @@ class ApiRequest {
     http.Response res;
     Uri uri = getRequestUri();
 
-    switch (method) {
-      case RequestMethod.PUT:
-        res = await http.put(uri, headers: headers, body: getBodyData());
-        break;
+    // if request have files
+    if (files.isNotEmpty && (method == RequestMethod.POST || method == RequestMethod.PUT)) {
+      http.MultipartRequest request = http.MultipartRequest(method == RequestMethod.PUT ? 'PUT' : 'POST', uri);
+      request.headers.addAll(headers);
+      files.forEach((key, file) {
+        request.files.add(http.MultipartFile.fromBytes(key, file.readAsBytesSync(), filename: basename(file.path)));
+      });
 
-      case RequestMethod.POST:
-        res = await http.post(uri, headers: headers, body: getBodyData());
-        break;
+      res = await http.Response.fromStream(await request.send());
+    } else {
+      // if request not have files
+      switch (method) {
+        case RequestMethod.PUT:
+          res = await http.put(uri, headers: headers, body: getBodyData());
+          break;
 
-      case RequestMethod.DELETE:
-        res = await http.delete(uri, headers: headers);
-        break;
+        case RequestMethod.POST:
+          res = await http.post(uri, headers: headers, body: getBodyData());
+          break;
 
-      case RequestMethod.GET:
-      default:
-        res = await http.get(uri, headers: headers);
-        break;
+        case RequestMethod.DELETE:
+          res = await http.delete(uri, headers: headers);
+          break;
+
+        case RequestMethod.GET:
+        default:
+          res = await http.get(uri, headers: headers);
+          break;
+      }
     }
 
     return ApiRequestResponse(res.statusCode, res.body);
